requirements.txt
asgiref==3.7.2
chardet==5.2.0
Django==5.0.1
mysqlclient==2.2.1
pillow==10.2.0
python-dateutil==2.8.2
reportlab==4.0.9
six==1.16.0
sqlparse==0.4.4
tzdata==2023.4
xlwt==1.3.0